feat(setup):  changes to docker script to keep the old data
	- The CP setup currently removes all the old rocksdb
	  made changes to support the old/new setup.

+++ scripts/docker/run-cpcontainer.sh
@@ -1,6 +1,9 @@
 #Provide start of the port range as command line arg
 #I.e sudo ./run-cpcontainer.sh 5000
-rm -rf *.raftdb configs ctl-interface logs
+if [ "$2" = "new" ]; then
+        rm -rf *.raftdb configs ctl-interface logs
+fi
+
 docker build -t controlplane .
 curdir=$(pwd)
-docker run -it -v "${curdir}:/controlplane" --network=host controlplane $1
+docker run -it -v "${curdir}:/controlplane" --network=host controlplane $1 $2 $3
